Trait iprange::IpNet

pub trait IpNet: ToNetwork<Self> + Debug + Ord + Copywhere
    Self: Sized,{
    type S: TraverseState<Net = Self>;
    type I: Iterator<Item = bool>;

    fn prefix_bits(&self) -> Self::I;
    fn prefix_len(&self) -> u8;
    fn with_new_prefix(&self, len: u8) -> Self;
}

An abstraction for IP networks.

Required Associated Types

type S: TraverseState<Net = Self>

Used for internal traversing.

type I: Iterator<Item = bool>

I is an iterator to the prefix bits of the network.

Required Methods

fn prefix_bits(&self) -> Self::I

Returns the iterator to the prefix bits of the network.

fn prefix_len(&self) -> u8

Returns the prefix length.

fn with_new_prefix(&self, len: u8) -> Self

Returns a copy of the network with the address truncated to the given length.